requirements.txt
alembic==1.6.5
bcrypt==3.2.0
cffi==1.14.6
click==7.1.2
colorama==0.4.4
dateutils==0.6.12
dnspython==2.1.0
email-validator==1.1.3
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-Migrate==3.1.0
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
flask-serialize==1.5.2
Flask-WTF==0.15.1
greenlet==1.1.0
idna==3.2
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
mongoengine==0.23.1
permissive-dict==1.0.4
pycparser==2.20
PyJWT==1.5.3
pymongo==3.10.1
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2021.1
six==1.14.0
Werkzeug==1.0.1
WTForms==2.3.3